apps/api/pyproject.toml
[build-system]
requires = ["setuptools==75.2.0", "wheel==0.44.0"]
build-backend = "setuptools.build_meta"

[project]
name = "omniflow-api"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi==0.131.0",
  "starlette==0.49.1",
  "uvicorn==0.32.0",
  "pydantic==2.9.2",
  "pydantic-settings==2.6.0",
  "sqlalchemy==2.0.36",
  "psycopg[binary]==3.2.3",
  "alembic==1.13.3",
  "redis==5.2.0",
  "celery==5.4.0",
  "pyyaml==6.0.2",
  "python-json-logger==2.0.7",
  "opentelemetry-api==1.28.2",
  "opentelemetry-sdk==1.28.2",
  "sentry-sdk==2.17.0"
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.3",
  "pytest-asyncio==0.24.0",
  "httpx==0.27.2",
  "ruff==0.7.1",
  "mypy==1.12.1",
  "types-redis==4.6.0.20241004",
  "types-PyYAML==6.0.12.20240917",
  "pip-audit==2.7.3"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["apps/api"]
markers = [
  "integration: tests requiring postgres and redis"
]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
